14:
* Extraction module player
* Division test load from multi raw
* Rename read raw yaml
* Fix tests
